refactor: 요청/응담 DTO 파일 구조 변경
