+++ src/main/java/com/umc/halo/domain/member/controller/MemberController.java
@@ -1,8 +1,8 @@
 package com.umc.halo.domain.member.controller;
 
 import com.umc.halo.domain.member.controller.docs.MemberControllerDocs;
-import com.umc.halo.domain.member.dto.request.LoginRequestDTO;
-import com.umc.halo.domain.member.dto.response.LoginResponseDTO;
+import com.umc.halo.domain.member.dto.MemberReqDTO;
+import com.umc.halo.domain.member.dto.MemberResDTO;
 import com.umc.halo.domain.member.exception.code.AuthSuccessCode;
 import com.umc.halo.domain.member.service.MemberService;
 import com.umc.halo.global.apiPayload.ApiResponse;
@@ -22,8 +22,8 @@ public class MemberController implements MemberControllerDocs {
     private final MemberService memberService;
 
     @PostMapping("/v1/auth/login")
-    public ApiResponse<LoginResponseDTO.LoginResponse> login(
-            @RequestBody @Valid LoginRequestDTO.Login dto
+    public ApiResponse<MemberResDTO.LoginResponse> login(
+            @RequestBody @Valid MemberReqDTO.Login dto
     ) {
         BaseSuccessCode code = AuthSuccessCode.LOGIN_SUCCESS;
         return ApiResponse.onSuccess(code, memberService.login(dto));
