+++ src/main/java/com/umc/halo/domain/member/converter/MemberConverter.java
@@ -1,6 +1,6 @@
 package com.umc.halo.domain.member.converter;
 
-import com.umc.halo.domain.member.dto.response.LoginResponseDTO;
+import com.umc.halo.domain.member.dto.MemberResDTO;
 import com.umc.halo.domain.member.entity.Member;
 import com.umc.halo.domain.member.enums.Provider;
 
@@ -13,13 +13,13 @@ public class MemberConverter {
                 .build();
     }
 
-    public static LoginResponseDTO.LoginResponse toLoginResponse(
+    public static MemberResDTO.LoginResponse toLoginResponse(
             String accessToken,
             String refreshToken,
             Boolean isNewUser,
             Boolean onboardingCompleted
     ) {
-        return LoginResponseDTO.LoginResponse.builder()
+        return MemberResDTO.LoginResponse.builder()
                 .accessToken(accessToken)
                 .refreshToken(refreshToken)
                 .isNewUser(isNewUser)
