+++ src/main/java/com/umc/halo/domain/member/dto/MemberResDTO.java
@@ -1,8 +1,8 @@
-package com.umc.halo.domain.member.dto.response;
+package com.umc.halo.domain.member.dto;
 
 import lombok.Builder;
 
-public class LoginResponseDTO {
+public class MemberResDTO {
 
     @Builder
     public record LoginResponse (
